Committing fixes for yahoo finance data pull, saving prevClose to the RPC server, prevent alerts from being overwritten, update rule file, and cleanup messy handling of bad alphavantage morning data.
ExtDataSources.py - Change yahoo finance api from v7 to v6 because this seems to work.
ProcessTopPercentGainers.py - Update process management so that once we send a BO alert, we don't send anymore for that stock.
                              Save prevClose to the RPC server.
StockProps.json             - Update target to 120%
                            - Set LAST_BUY_TIME to 12:30
UpdateAfternoonBreakouts.py - Clean up the handling of the tHOD_Initial set since the alphavantage data sometimes is missing the first few minutes and it was a clunky implementation before.