Make some updates in preparation to share this repository publicly.
ExtDataSourcesOriginal.py - Unused in this repo so getting rid of this file.
Communicator.py           - Clean up contacts list.
ExtDataSources.py         - Remove unused sendMessage and contacts definitions from this file.  It used to host this functionality, but that was moved to Communicator.
IBDConnParams.json        - Remove strategy definitions unrelated to afternoon breakouts.
README.md                 - Provide an explanation for the repository files and which files work and which still need some development.